+++ src/Symfony/Component/PropertyInfo/Tests/Fixtures/FalseAccessorDummy.php
@@ -1,0 +1,49 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\PropertyInfo\Tests\Fixtures;
+
+class FalseAccessorDummy
+{
+    private $Foo;
+
+    public function hash(): void
+    {
+    }
+
+    public function cancel()
+    {
+    }
+
+    public function gettings()
+    {
+    }
+
+    public function settings()
+    {
+    }
+
+    public function isolate()
+    {
+    }
+
+    public function getValid()
+    {
+    }
+
+    public function getFoo()
+    {
+    }
+
+    public function get_foo()
+    {
+    }
+}